Add a CLI argument that allows to specify a prerelease prefix (#459)
SemVer specification allows prerelease version to be a string/number combination. For example: 1.0.0-beta.5, where beta.5 is a prerelease version. Unfortunately, beachball does not give a user control over the prerelease prefix. This PR addresses this at the level of the bump command.

This is extremely useful if you want to scope versions to avoid conflicts. For example, 1.0.0-beta.2 and 1.0.0-alpha.2 are two different versions. Without --prerelease-prefix option, beachball will be calculating the same version (1.0.0-2 and 1.0.0-2) for two packages that have different contents.