Fix setting response.type in promptForChange
This commit handles edge case for properly setting `response.type` in
`promptForChange` in a situation when `options.type` is `undefined` or
`null` **and** there is only one viable default `ChangeType` option
(`'none'`) in which case the type prompt won't be shown, and we
fall back to `'none'` by default.

Prior to this change, since `options.type` would be left `undefined`
and the type prompt not shown, the CLI would incorrectly exit with
"Prompt response contains invalid change type" without any other hint
to the user.