Allow passing --config to specify config file (#477)
* Allow passing --config to specify config file

Automated testing for this change was a bit tricky. To make options code more testable, we allowing injecting argv instead of grabbing from the process, which we should not mutate, and may have flags we don't want. E.g. E2ETest will run jest with a `--config` flag we don't want in our tests.

This exposed some issues where `getCliOptions` was accidentally trimming bits off of process.argv, and we needed to change the CLI options cache to allow different values of argv to have different results. We do this by special casing the cache to only cache when the input is process.argv, which should be immutable and will be the only optyion used in production code.

* Change files

* Inject CLI options into getRootOptions instead of argv