chore(BeachballOptions): Move `scope` to `RepoOptions` type
The `scope` CLI option, can be used in the configuration file regardless
of it being a cli specific option.

Adds `scope` to the `RepoOptions` type so that it can be used in the
config file as a part of `BeachballConfig`

Also improves the options types by making `CliOptions` extend shared
`RepoOptions`.