feat: Introduce --fixup to change subcommand
When working on a change in a monorepo, I will often need to add
subsequent change files as the change spreads into multiple packages.
This creates multiple changefiles as well as multiple commits.

To tackle this, this change introduces a `--fixup` flag, modeled after
the `--fixup` flag in `git commit`. When used, this will enable fix-up
mode, which will:

 1. Find the most recent commit which adds a change file
 2. Add the new changes into that file
 3. Commit with a `--fixup <sha>` flag

If no commits are found in the current branch, then the standard flow is
used to create a new commit / changefile.